refactor: group error/loading redux files
