+++ src/reducers/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux'
 import { routerReducer } from 'react-router-redux'
-import loading from './loadingReducer'
-import error from './errorReducer'
+import loading from '../store/loading/loadingReducer'
+import error from '../store/error/errorReducer'
 import date from '../store/date/dateReducer'
 import items from '../store/expenses/expensesReducer'
 import authentication from '../store/auth/authReducer'
