+++ src/scenes/addExpense/index.js
@@ -2,7 +2,7 @@ import { push } from 'react-router-redux'
 import { connect } from 'react-redux'
 import { addItem } from '../../store/expenses/expensesActions'
 import AddExpense from './AddExpense'
-import { createLoadingSelector } from '../../selectors/loadingSelectors'
+import { createLoadingSelector } from '../../store/loading/loadingSelectors'
 
 const mapStateToProps = state => ({
   isPosting: createLoadingSelector(['ADD_ITEM'])(state),
