+++ src/scenes/login/index.js
@@ -1,8 +1,8 @@
 import { connect } from 'react-redux'
 import LoginForm from './LoginForm'
 import { getCurrentSession, signIn } from '../../store/auth/authActions'
-import { createLoadingSelector } from '../../selectors/loadingSelectors'
-import { createErrorMessageSelector } from '../../selectors/errorSelectors'
+import { createLoadingSelector } from '../../store/loading/loadingSelectors'
+import { createErrorMessageSelector } from '../../store/error/errorSelectors'
 
 const mapStateToProps = state => ({
   authentication: state.authentication,
